requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
idna==3.4
pip==21.3.1
pydantic-core==2.10.1
pydantic==2.4.2
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
setuptools==60.2.0
sniffio==1.3.0
starlette==0.31.1
typing-extensions==4.8.0
urllib3==2.0.5
wheel==0.37.1
